Add slides for Projectsveltos
Projectsveltos is a Kubernetes add-on controller that makes it easy
to deploy and manage add-ons across multiple clusters.
It supports a variety of add-on formats, including Helm charts,
raw YAMLs, Kustomize, Carvel ytt, and Jsonnet.

Projectsveltos uses templates to represent add-ons, which can then
be instantiated on each cluster during deployment. This allows you
to use the same add-on configuration across all of your clusters,
while still allowing for some variation, such as different add-on
configuration values.

Projectsveltos can get the information it needs to instantiate the templates
from either the management cluster or the managed clusters themselves.
This flexibility makes Sveltos a powerful tool for managing add-ons in a
variety of environments.